feat: modularised configuration, file was too big
